% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_year.R
\name{add_year}
\alias{add_year}
\title{Add year column in data frame}
\usage{
add_year(data, year)
}
\arguments{
\item{data}{data frame}

\item{year}{vector}
}
\description{
Add year column in data frame
}
\examples{
\dontrun{df |> add_year(2001:2020)}
}
