% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_cell.R
\name{pt_cell}
\alias{pt_cell}
\title{Split long vector to short}
\usage{
pt_cell(data)
}
\arguments{
\item{data}{vector data, extracted from pdf, mostly character.}
}
\value{
splited data, each element has a each record.
}
\description{
Split long vector to short
}
\examples{
\dontrun{conbini |> pt_cell()}
}
