% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_conbini.R
\name{pt_conbini}
\alias{pt_conbini}
\title{Fixing character to numeric data}
\usage{
pt_conbini(data)
}
\arguments{
\item{data}{vector or list of vector}
}
\description{
Fixing character to numeric data
}
\examples{
\dontrun{df |> pt_split_multi() |> pt_conbini()}
}
