% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_index.R
\name{pt_index}
\alias{pt_index}
\title{Create index for extracted data from pdf.}
\usage{
pt_index(data, keywords)
}
\arguments{
\item{data}{vector or list of vector}

\item{keywords}{character vector which indicates first column of each raw.}
}
\description{
Create index for extracted data from pdf.
}
\examples{
\dontrun{df |> pt_index()}
}
