% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_modify.R
\name{pt_modify}
\alias{pt_modify}
\title{Fixing character to numeric data}
\usage{
pt_modify(data)
}
\arguments{
\item{data}{vector or list of vector}
}
\description{
Fixing character to numeric data
}
\examples{
\dontrun{df |> pt_split_multi() |> pt_modify()}
}
