% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_split.R
\name{pt_split}
\alias{pt_split}
\title{Extract data}
\usage{
pt_split(data, initial, n_col = 15, dur_month = 1)
}
\arguments{
\item{data}{vector or list of vector}

\item{initial}{number of begining}

\item{n_col}{number of columns}

\item{dur_month}{duration in months}
}
\description{
Extract data
}
\examples{
\dontrun{
conbini |>
  pt_cell() |>
  purrr::flatten_chr() |>
  pt_split(initial = 29)
  }
}
