% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_split_multi.R
\name{pt_split_multi}
\alias{pt_split_multi}
\title{Extract data}
\usage{
pt_split_multi(data, index = idx)
}
\arguments{
\item{data}{vector}

\item{index}{data frame}
}
\description{
Extract data
}
\examples{
\dontrun{df |> pt_split_multi()}
}
