% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_split_one.R
\name{pt_split_one}
\alias{pt_split_one}
\title{Extract data}
\usage{
pt_split_one(data, initial, n_col = 15)
}
\arguments{
\item{data}{vector or list of vector}

\item{initial}{number}

\item{n_col}{number of columns}
}
\description{
Extract data
}
\examples{
\dontrun{
conbini |>
  pt_cell() |>
  purrr::flatten_chr() |>
  pt_split_one(initial = idx[12,1])
  }
}
