% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_split_year.R
\name{pt_split_year}
\alias{pt_split_year}
\title{Extract data}
\usage{
pt_split_year(data, index = idx[seq(nrow(idx)), 1], n_col = 15)
}
\arguments{
\item{data}{vector or list of vector}

\item{index}{vector which passes into initial number,
appropriate index is required.}

\item{n_col}{number of columns}
}
\description{
Extract data
}
\examples{
conbini |>
  pt_cell() |>
  purrr::flatten_chr() |>
  pt_split_year()
}
